Modifications to the Location work

+++ Models/Website.cs
@@ -5,10 +5,10 @@ using System.Web;
 
 namespace GCRBA.Models {
 	public class Website {
-		public Int16 intWebsiteID = 0;
+		public long intWebsiteID = 0;
 		public long intCompanyID = 0;
 		public string strURL = string.Empty;
-		public Int16 intWebsiteTypeID = 0;
+		public short intWebsiteTypeID = 0;
 		public string strWebsiteType = string.Empty;
 	}
 }
